pyproject.toml
[build-system]
requires = ["setuptools<69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devsync"
version = "0.11.0"
description = "Distribute and sync dev tool configurations across teams"
readme = "README.md"
authors = [{ name = "Troy Larson", email = "troy@calvinware.com" }]
requires-python = ">=3.10"
license = { text = "MIT License" }
keywords = ["cli", "ai", "config", "mcp", "cursor", "copilot", "claude", "cline", "codex", "kiro", "roo", "windsurf"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "typer[all]>=0.15.0",
    "rich>=14.0.0",
    "pyyaml>=6.0",
    "textual>=6.0.0",
    "GitPython>=3.1.45",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "black==24.10.0",
    "mypy>=1.10.0",
    "ruff>=0.9.0",
    "invoke>=2.0.0",
    "build>=1.2.1",
    "twine>=5.0.0",
    "types-PyYAML>=6.0.12.20240808",
]

[project.scripts]
devsync = "devsync.cli.main:app"

[project.urls]
Homepage = "https://github.com/troylar/devsync"
Documentation = "https://devsync.readthedocs.io"
Repository = "https://github.com/troylar/devsync"
Issues = "https://github.com/troylar/devsync/issues"

[tool.setuptools.packages.find]
include = ["devsync*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=devsync --cov-report=term-missing -p no:cacheprovider"

[tool.coverage.run]
relative_files = true
source = ["devsync"]

[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312", "py313"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
